requirements.txt
Flask==1.0.2
flask-sqlalchemy==2.4.0
Flask-User==1.0.1.5
pyscss==1.3.5

tinytag==1.2.2
pylast==3.1.0

celery==4.3.0
redis==3.2.1
Flask-Caching==1.7.2
